Moved public globals in lib.rs Clearer separation of code between files
