+++ wasm-test/src/lib.rs
@@ -3,7 +3,6 @@ extern crate hdk;
 extern crate holochain_wasm_utils;
 
 use holochain_wasm_utils::*;
-use hdk::globals;
 use hdk::globals::g_mem_stack;
 
 #[no_mangle]
@@ -11,12 +10,12 @@ pub extern "C" fn check_global_dispatch(encoded_allocation_of_input : i32) -> i3
   unsafe {
     g_mem_stack = Some(SinglePageStack::new_from_encoded(encoded_allocation_of_input as u32));
 
-    hdk::debug(&globals::APP_NAME);
-    hdk::debug(&globals::APP_DNA_HASH);
-    hdk::debug(&globals::APP_KEY_HASH);
-    hdk::debug(&globals::APP_AGENT_HASH);
-    hdk::debug(&globals::APP_AGENT_TOP_HASH);
-    hdk::debug(&globals::APP_AGENT_STR);
+    hdk::debug(&hdk::APP_NAME);
+    hdk::debug(&hdk::APP_DNA_HASH);
+    hdk::debug(&hdk::APP_KEY_HASH);
+    hdk::debug(&hdk::APP_AGENT_HASH);
+    hdk::debug(&hdk::APP_AGENT_TOP_HASH);
+    hdk::debug(&hdk::APP_AGENT_STR);
   }
   return 0;
 }
